% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PIHM.R
\name{PIHM}
\alias{PIHM}
\title{A Cat Function}
\usage{
PIHM(param = TRUE)
}
\arguments{
\item{Love}{PIHM ?Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
PIHM()
}
\keyword{PIHM}

