% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PIHM.R
\name{PIHM}
\alias{PIHM}
\title{PIHM Analysis project.
Developed by Lele Shu( lele.shu at gmail.com  lzs157 at psu.edu )}
\usage{
PIHM(param = TRUE)
}
\arguments{
\item{Love}{PIHM? Defaults to TRUE.}
}
\description{
=============================================
 Current version is for PIHM-MF or PIHM v2.4;
}
\details{


=============================================
}
\examples{
PIHM()
}
\keyword{PIHM}

